+++ scriptofan/src/main/java/com/springserver/api/repository/TransactionRepository.java
@@ -4,6 +4,8 @@ import com.springserver.api.model.Transaction;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface TransactionRepository extends CrudRepository<Transaction, String> {
 }
